Read initial pong count from db

+++ ping-pong/models/counter.go
@@ -1,5 +1,5 @@
 package models
 
 type Counter struct {
-	PingCount int64
+	PingCount int
 }
